FINI

+++ VerifCoin.m
@@ -1,6 +1,6 @@
 function [] = VerifCoin(image,coins,video)
 % on affiche l'image 
-imshow(image);
+imshow(uint8(image));
 hold on;
 % on affiche les coinns 
 for i=1:size(coins,1)
